hello_activemq/requirements.txt
django >= 2.1.2
stomp.py >= 5.0.1
